requirements.txt
Django==3.0.3
djangorestframework==3.11.0
Markdown>=3.0.1